Rename plausibleUtils → analyticsUtils
