+++ src/routes/+page.svelte
@@ -22,7 +22,7 @@
 	import { getUserId } from '$lib/utils/user';
 	import analytics from '$lib/Analytics';
 	import { userLocation } from '$src/stores/userLocationStore';
-	import { analyticsDistanceToStop } from '$lib/Analytics/plausibleUtils';
+	import { analyticsDistanceToStop } from '$lib/Analytics/analyticsUtils';
 	import SurveyLauncher from '$components/surveys/SurveyLauncher.svelte';
 	import { parseInitialCoordinates, cleanUrlParams } from '$lib/urlParams';
 	import TripOptionsModal from '$components/trip-planner/TripOptionsModal.svelte';
