+++ src/routes/stops/[stopID]/+page.svelte
@@ -8,7 +8,7 @@
 	import { loadSurveys } from '$lib/Surveys/surveyUtils.js';
 	import { getUserId } from '$lib/utils/user.js';
 	import analytics from '$lib/Analytics';
-	import { analyticsDistanceToStop } from '$lib/Analytics/plausibleUtils.js';
+	import { analyticsDistanceToStop } from '$lib/Analytics/analyticsUtils.js';
 	import { userLocation } from '$src/stores/userLocationStore.js';
 
 	let { data } = $props();
